+++ Pods/FirebaseAuth/Firebase/Auth/Source/Public/FIRAuthErrors.h
@@ -304,6 +304,11 @@ typedef NS_ENUM(NSInteger, FIRAuthErrorCode) {
      */
     FIRAuthErrorCodeNullUser = 17067,
 
+    /** Indicates that the Firebase Dynamic Link domain used is either not configured or is unauthorized
+        for the current project.
+     */
+    FIRAuthErrorCodeInvalidDynamicLinkDomain = 17074,
+
     /** Indicates an error occurred while attempting to access the keychain.
      */
     FIRAuthErrorCodeKeychainError = 17995,
@@ -311,6 +316,11 @@ typedef NS_ENUM(NSInteger, FIRAuthErrorCode) {
     /** Indicates an internal error occurred.
      */
     FIRAuthErrorCodeInternalError = 17999,
+
+    /** Raised when a JWT fails to parse correctly. May be accompanied by an underlying error
+        describing which step of the JWT parsing process failed.
+     */
+    FIRAuthErrorCodeMalformedJWT = 18000,
 } NS_SWIFT_NAME(AuthErrorCode);
 
 @end