+++ Pods/FirebaseAuth/Firebase/Auth/Source/RPCs/FIRAuthBackend.m
@@ -284,6 +284,12 @@ static NSString *const kMissingAndroidPackageNameErrorMessage = @"MISSING_ANDROI
  */
 static NSString *const kUnauthorizedDomainErrorMessage = @"UNAUTHORIZED_DOMAIN";
 
+/** @var kInvalidDynamicLinkDomainErrorMessage
+ @brief This is the error message the server will respond with if the dynamic link domain provided
+ in the request is invalid.
+ */
+static NSString *const kInvalidDynamicLinkDomainErrorMessage = @"INVALID_DYNAMIC_LINK_DOMAIN";
+
 /** @var kInvalidContinueURIErrorMessage
     @brief This is the error message the server will respond with if the continue URL provided in
         the request is invalid.
@@ -516,7 +522,7 @@ static id<FIRAuthBackendImplementation> gBackendImplementation;
 
 - (void)asyncPostToURLWithRequestConfiguration:(FIRAuthRequestConfiguration *)requestConfiguration
                                            URL:(NSURL *)URL
-                                          body:(NSData *)body
+                                          body:(nullable NSData *)body
                                    contentType:(NSString *)contentType
                              completionHandler:(void (^)(NSData *_Nullable,
                                                          NSError *_Nullable))handler {
@@ -1042,6 +1048,10 @@ static id<FIRAuthBackendImplementation> gBackendImplementation;
 
   if ([shortErrorMessage isEqualToString:kInvalidContinueURIErrorMessage]) {
     return [FIRAuthErrorUtils invalidContinueURIErrorWithMessage:serverDetailErrorMessage];
+  }
+
+  if ([shortErrorMessage isEqualToString:kInvalidDynamicLinkDomainErrorMessage]) {
+    return [FIRAuthErrorUtils invalidDynamicLinkDomainErrorWithMessage:serverDetailErrorMessage];
   }
 
   if ([shortErrorMessage isEqualToString:kMissingContinueURIErrorMessage]) {