+++ Pods/FirebaseAuth/Firebase/Auth/CHANGELOG.md
@@ -1,3 +1,14 @@
+# v5.0.5
+- Restore SafariServices framework dependency (#2002).
+
+# v5.0.4
+- Fix analyzer issues (#1740).
+
+# v5.0.3
+- Adds `FIRAuthErrorCodeMalformedJWT`, which is raised on JWT token parsing.
+  failures during auth operations (#1436).
+- Migrate to use FirebaseAuthInterop interfaces to access FirebaseAuth (#1501).
+
 # v5.0.2
 - Fix an issue where JWT date timestamps weren't parsed correctly. (#1319)
 - Fix an issue where anonymous accounts weren't correctly promoted to